// src/app/models/Abono.ts
export class Abono {

    id: string | undefined;
    id_prestamo: string | undefined;
    Supervisor: string | undefined;
    id_supervisor: string | undefined;
    Cliente: string | undefined;
    id_cliente: string| undefined;
    Promotor: string | undefined;
    id_promotor: string| undefined;
    fecha_abono: string| undefined;
    cantidad_abonar: number| undefined;
    total_pagado: number| undefined;
    cambio: number| undefined;
    debia: number| undefined;
    ahora_debe: number| undefined;
    created_at: string| undefined;
}